// src_migration/types.ts
export interface Product {
  id: number;
  nome: string;
  codigo_marca?: string | number;
  marca: string;
  sabor?: string;
  peso?: string;
  foto: string;
  infonutri?: string;
  preco: number;
  precoNaoPromocional?: number;
  precoOriginal?: number;
  precoMedio?: number;
  nota?: number | string;
  categoria?: string[];
  variacao?: number[];
  desconto?: number;
  comentario?: string;
  disponivel: boolean | number;
  palavrasChave?: string[];
}
// ...
export interface CheckoutProduct {
  id: number;
  quantidade: number;
  preco: number;
  precoFinal: number;
  codigo_marca?: string | number;
}